+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## DEVEL
 ### New Features
+- synchronized geocoding types with current geocoding API
+
 ### Bug Fixes
 ### Others